+++ dipu/torch_dipu/csrc_dipu/profiler/profiler_kineto.cpp
@@ -6,10 +6,6 @@
 #include <variant>
 
 #include <c10/macros/Export.h>
-#if DIPU_TORCH_VERSION >= 20200
-#include <c10/util/ApproximateClock.h>
-#endif
-#include <c10/util/C++17.h>
 #include <c10/util/Exception.h>
 #include <c10/util/flat_hash_map.h>
 #include <c10/util/irange.h>
@@ -25,31 +21,12 @@
 #include "csrc_dipu/utils/Log.h"
 
 #include "collection.h"
+#include "torch_version.h"
 
 namespace dipu {
 namespace profile {
 
-#if DIPU_TORCH_VERSION >= 20200
-using c10::ApproximateClockToUnixTimeConverter;
-using c10::getApproximateTime;
-using std::holds_alternative;
-namespace variant = std;
-#else
-using c10::holds_alternative;
-using torch::profiler::impl::ApproximateClockToUnixTimeConverter;
-using torch::profiler::impl::getApproximateTime;
-namespace variant = c10;
-#endif
-
 namespace {
-inline int64_t getTimeUs() {
-  auto constexpr scale = int64_t{1000};
-#if DIPU_TORCH_VERSION >= 20200
-  return c10::getTime() / scale;
-#else
-  return torch::profiler::impl::getTime() / scale;
-#endif
-}
 
 const std::set<libkineto::ActivityType> kCpuTypes {
   libkineto::ActivityType::CPU_OP, libkineto::ActivityType::CPU_INSTANT_EVENT,
@@ -105,19 +82,20 @@ auto parseArgData(const std::vector<op_input_t>& input_shapes,
   std::vector<c10::IValue> concrete_inputs_list;
 
   for (const auto& i : c10::irange(input_shapes.size())) {
-    variant::visit(c10::overloaded(
-                       [&](const TensorMetadata& t) {
-                         shapes[i] = t.sizes_;
-                         dtypes[i] = std::string(
-                             scalarTypeToTypeMeta(t.dtype_).name().data(),
-                             scalarTypeToTypeMeta(t.dtype_).name().size());
-                       },
-                       [&](const std::vector<TensorMetadata>&) {
-                         dtypes[i] = "TensorList";
-                       },
-                       [&](const c10::IValue& val) { dtypes[i] = "Scalar"; },
-                       [&](const auto&) {}),
-                   input_shapes[i]);
+    dipu::profile::visit(
+        c10::overloaded(
+            [&](const TensorMetadata& t) {
+              shapes[i] = t.sizes_;
+              dtypes[i] =
+                  std::string(scalarTypeToTypeMeta(t.dtype_).name().data(),
+                              scalarTypeToTypeMeta(t.dtype_).name().size());
+            },
+            [&](const std::vector<TensorMetadata>&) {
+              dtypes[i] = "TensorList";
+            },
+            [&](const c10::IValue& val) { dtypes[i] = "Scalar"; },
+            [&](const auto&) {}),
+        input_shapes[i]);
   }
 
   // If we recorded concrete inputs, then parse them
@@ -126,18 +104,18 @@ auto parseArgData(const std::vector<op_input_t>& input_shapes,
     concrete_inputs_list.resize(input_shapes.size());
 
     for (const auto& i : c10::irange(input_shapes.size())) {
-      variant::visit(
+      dipu::profile::visit(
           c10::overloaded(
               [&](const c10::IValue& val) { concrete_inputs_list[i] = val; },
               [&](const auto&) {}),
           input_shapes[i]);
-      variant::visit(c10::overloaded(
-                         [&](const c10::IValue& val) {
-                           concrete_inputs_list[i] = val;
-                           dtypes[i] = "ScalarList";
-                         },
-                         [&](const auto&) {}),
-                     concrete_inputs[i]);
+      dipu::profile::visit(c10::overloaded(
+                               [&](const c10::IValue& val) {
+                                 concrete_inputs_list[i] = val;
+                                 dtypes[i] = "ScalarList";
+                               },
+                               [&](const auto&) {}),
+                           concrete_inputs[i]);
     }
   }
